Refactored TaskDetails
