--- frontend/src/components/taskDetailsHeader/TaskDetailsHeader.css
@@ -1,5 +1,0 @@
-.task-details-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
