--- frontend/src/components/taskDetailsRest/TaskDetailsRest.css
@@ -1,86 +1,0 @@
-.other-details {
-  display: flex;
-  flex-direction: column;
-  gap: 0.8rem;
-}
-
-.single-detail {
-  display: flex;
-  flex-direction: column;
-  gap: 0.6rem;
-}
-
-.single-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.attachment-files {
-  display: flex;
-  align-items: center;
-  gap: 0.8rem;
-}
-
-.attachment-files img {
-  width: 4rem;
-  height: 4rem;
-  border-radius: 8px;
-}
-
-#small-check {
-  width: 1.2rem;
-  height: 1.2rem;
-  color: #5fb918;
-}
-
-#small-circle {
-  width: 1.2rem;
-  height: 1.2rem;
-  color: #b3b1bb;
-}
-
-.objects {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.object,
-.comment-section {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.object p,
-.comment-section p {
-  font-family: 'Manrope', sans-serif;
-  font-size: 14px;
-  font-weight: 500;
-  line-height: 21px;
-  color: #2b283d;
-}
-
-.comment-part:first-child {
-  color: #af20c7;
-}
-
-.question {
-  background: transparent;
-  width: 100%;
-  height: 3rem;
-  border: 1px solid #d8e0e8;
-  border-radius: 8px;
-  padding: 0.8rem;
-  color: #b3b1bb;
-}
-
-.comment-text {
-  font-size: 14px;
-  font-family: 'Manrope', sans-serif;
-  font-weight: 500;
-  line-height: 21px;
-  color: #4c4958;
-  padding-left: 2rem;
-}
