+++ frontend/src/pages/dashboard/Dashboard.tsx
@@ -8,7 +8,7 @@ import { signOut, User, getAuth, onAuthStateChanged } from '@firebase/auth';
 import api from '../../api';
 
 const Dashboard = () => {
-  const [loading, setLoading] = useState<boolean>(true);
+  const [loading, setLoading] = useState<boolean>(false);
 
   const navigate = useNavigate();
   const handleSignOut = () => signOut(getAuth());
@@ -30,9 +30,9 @@ const Dashboard = () => {
     });
   };
 
-  useEffect(() => {
-    checkToken();
-  }, []);
+  // useEffect(() => {
+  //   checkToken();
+  // }, []);
 
   return (
     <>
